+++ src/utils/get-file-content-or-null.js
